backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai-whisper==20240930
numpy==2.2.0
scipy==1.15.0
torch==2.5.0
pydantic==2.10.0
python-dotenv==1.0.1
